Move PositionalEncoding from core_validator to neuron_utilities
