+++ bittensor/_neuron/text/core_validator/__init__.py
@@ -546,40 +546,6 @@ class neuron:
                     del self.neuron_stats[uid]
 
 
-class PositionalEncoding(nn.Module):
-    r""" Positional Encoder which adds information based on the relative position of each token
-    
-    """
-    def __init__(self, d_model: int, dropout: float, max_len: int = 5000):
-        super().__init__()
-        self.dropout = nn.Dropout(p=dropout)
-
-        position = torch.arange(max_len).unsqueeze(1)
-        div_term = torch.exp(torch.arange(0, d_model, 2) * (-math.log(10000.0) / d_model))
-
-        # === Create position matrix ===
-        # Creates a positional matrix with alternating frequencies 
-        # pe: (torch.FloatTensor) positional encoding matrix
-        # pe.shape: [1, max_len, network_dim]
-        pe = torch.zeros(1, max_len, d_model)
-        pe[0, :, 0::2] = torch.sin(position * div_term)
-        pe[0, : , 1::2] = torch.cos(position * div_term)
-        self.register_buffer('pe', pe)
-
-    def forward(self, x: torch.tensor) -> torch.tensor:
-        """
-        Args:
-            x: Tensor, shape [batch_size, seq_len, embedding_dim]
-        """
-        # === Positional Encoding ===
-        # Inject some information of the relative position of the token in the sequence.
-        #  Finally, Dropout is applied to tokens
-        # x: (torch.FloatTensor) input sequence tokens with position information injected
-        # x.shape: [batch_size, seq_len, network_dim]
-        x = x + self.pe[0, :x.size(1)]
-        return self.dropout(x)
-
-
 class nucleus( torch.nn.Module ):
     """ Nucleus class which holds the validator model.
     """
